Refactor aidants_connect to put common code into common package
