+++ aidants_connect_habilitation/apps.py
@@ -5,4 +5,4 @@ class AidantConnectHabilitationConfig(AppConfig):
     name = "aidants_connect_habilitation"
 
     def ready(self):
-        import aidants_connect.lookups  # noqa
+        import aidants_connect.common.lookups  # noqa
