+++ aidants_connect_habilitation/forms.py
@@ -4,8 +4,8 @@ from django.forms import formset_factory, ChoiceField, CharField, BaseFormSet
 from phonenumber_field.formfields import PhoneNumberField
 from phonenumber_field.widgets import PhoneNumberInternationalFallbackWidget
 
-from aidants_connect.constants import RequestOriginConstants
-from aidants_connect.forms import PatchedErrorListForm, PatchedErrorList
+from aidants_connect.common.constants import RequestOriginConstants
+from aidants_connect.common.forms import PatchedErrorListForm, PatchedErrorList
 from aidants_connect_habilitation import models
 from aidants_connect_habilitation.models import AidantRequest
 from aidants_connect_web.models import OrganisationType
