+++ aidants_connect_habilitation/tests/test_models.py
@@ -1,7 +1,7 @@
 from django.db import IntegrityError
 from django.test import tag, TestCase
 
-from aidants_connect.constants import RequestOriginConstants
+from aidants_connect.common.constants import RequestOriginConstants
 from aidants_connect_habilitation.tests.factories import OrganisationRequestFactory
