+++ aidants_connect_web/apps.py
@@ -6,4 +6,4 @@ class AidantConnectWebConfig(AppConfig):
 
     def ready(self):
         import aidants_connect_web.signals  # noqa
-        import aidants_connect.lookups  # noqa
+        import aidants_connect.common.lookups  # noqa
