+++ aidants_connect_web/tests/test_functional/test_a11y.py
@@ -1,6 +1,6 @@
 from django.test import tag
 
-from aidants_connect_web.tests.test_functional.testcases import FunctionalTestCase
+from aidants_connect.common.tests.testcases import FunctionalTestCase
 from selenium.common.exceptions import NoSuchElementException
