+++ aidants_connect_web/tests/test_functional/test_cancel_autorisation.py
@@ -11,7 +11,7 @@ from aidants_connect_web.tests.factories import (
     AutorisationFactory,
     UsagerFactory,
 )
-from aidants_connect_web.tests.test_functional.testcases import FunctionalTestCase
+from aidants_connect.common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.tests.test_functional.utilities import login_aidant
