+++ aidants_connect_web/tests/test_functional/test_cancel_mandat.py
@@ -8,7 +8,7 @@ from aidants_connect_web.tests.factories import (
     MandatFactory,
     AutorisationFactory,
 )
-from aidants_connect_web.tests.test_functional.testcases import FunctionalTestCase
+from aidants_connect.common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.tests.test_functional.utilities import login_aidant
