+++ aidants_connect_web/tests/test_functional/test_create_mandat.py
@@ -7,7 +7,7 @@ from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.support.expected_conditions import url_matches
 
 from aidants_connect_web.tests.factories import AidantFactory
-from aidants_connect_web.tests.test_functional.testcases import FunctionalTestCase
+from aidants_connect.common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.tests.test_functional.utilities import login_aidant
