+++ aidants_connect_web/tests/test_functional/test_import_aidant.py
@@ -13,7 +13,7 @@ from aidants_connect_web.tests.factories import (
     OrganisationFactory,
     CarteTOTPFactory,
 )
-from aidants_connect_web.tests.test_functional.testcases import FunctionalTestCase
+from aidants_connect.common.tests.testcases import FunctionalTestCase
 
 
 @tag("functional", "import")
