+++ aidants_connect_web/tests/test_functional/test_statistiques.py
@@ -1,6 +1,6 @@
 from django.test import tag
 
-from aidants_connect_web.tests.test_functional.testcases import FunctionalTestCase
+from aidants_connect.common.tests.testcases import FunctionalTestCase
 
 
 @tag("functional")
