+++ aidants_connect_web/tests/test_functional/test_usagers.py
@@ -4,7 +4,7 @@ from aidants_connect_web.tests.factories import (
     MandatFactory,
     ExpiredMandatFactory,
 )
-from aidants_connect_web.tests.test_functional.testcases import FunctionalTestCase
+from aidants_connect.common.tests.testcases import FunctionalTestCase
 from aidants_connect_web.tests.test_functional.utilities import login_aidant
