+++ aidants_connect_web/tests/test_views/test_espace_aidant/test_espace_aidant.py
@@ -5,7 +5,7 @@ from django.urls import resolve, reverse
 
 from django_otp.plugins.otp_totp.models import TOTPDevice
 
-from aidants_connect.constants import JournalActionKeywords
+from aidants_connect.common.constants import JournalActionKeywords
 from aidants_connect_web.models import Journal
 from aidants_connect_web.tests.factories import (
     AidantFactory,
